startup_investment_analyst/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "startup_investment_analyst"
version = "0.1.0"
description = "Custom agent tools and sub-agents for investment workflow"
authors = [
  {name="Your Name", email="your.email@example.com"}
]
dependencies = [
    "google-cloud-aiplatform[adk,agent_engines]",
    "google-cloud-bigquery",
    "google-cloud-storage",
    "cloudpickle==3.1.1",
    "pydantic==2.11.7",
    "beautifulsoup4",
    "requests"
]

[tool.setuptools.packages.find]
where = ["."]
